# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/checkins": {
            "get": {
                "description": "Returns all check-ins for the authenticated user, newest first. JWT required.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "checkin"
                ],
                "summary": "Get user's check-in history",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CheckinResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "User submits daily check-in with location and optional GPS. Only one per day. JWT required. If late, must provide reason.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "checkin"
                ],
                "summary": "Submit daily check-in",
                "parameters": [
                    {
                        "description": "Check-in data",
                        "name": "checkin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CheckinRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CheckinResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/checkins/all": {
            "get": {
                "description": "HR/admin only. Returns paginated list of all check-ins. Query params: page, page_size, user_id, date",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "checkin"
                ],
                "summary": "List all check-ins (HR/admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 20)",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by user ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by date (YYYY-MM-DD)",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CheckinResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/checkins/today": {
            "get": {
                "description": "Returns today's check-in for the authenticated user, or 404 if none. JWT required.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "checkin"
                ],
                "summary": "Get today's check-in",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CheckinResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/checkins/{id}": {
            "put": {
                "description": "Users cannot update check-ins. Use HR endpoint.",
                "tags": [
                    "checkin"
                ],
                "summary": "Not allowed",
                "responses": {
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "HR/admin only. Soft delete a check-in by setting deleted=true.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "checkin"
                ],
                "summary": "Soft delete check-in (HR/admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Check-in ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/dashboard/checkins/{id}": {
            "put": {
                "description": "HR or admin can update any user's check-in. Audit log is written. JWT with hr/admin role required.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "HR/admin update check-in",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Check-in ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Check-in data",
                        "name": "checkin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CheckinRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CheckinResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CheckinRequest": {
            "type": "object",
            "required": [
                "date",
                "location_type"
            ],
            "properties": {
                "date": {
                    "type": "string"
                },
                "gps_lat": {
                    "type": "number"
                },
                "gps_long": {
                    "type": "number"
                },
                "late_reason": {
                    "type": "string"
                },
                "location_detail": {
                    "type": "string"
                },
                "location_type": {
                    "type": "string",
                    "enum": [
                        "home",
                        "office",
                        "client",
                        "temporary"
                    ]
                },
                "notes": {
                    "type": "string"
                },
                "time": {
                    "description": "still accept string for backward compatibility",
                    "type": "string"
                }
            }
        },
        "models.CheckinResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "gps_lat": {
                    "type": "number"
                },
                "gps_long": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "late": {
                    "type": "boolean"
                },
                "late_reason": {
                    "type": "string"
                },
                "location_detail": {
                    "type": "string"
                },
                "location_type": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "time": {
                    "description": "return as RFC3339 string for API clients",
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string",
                    "example": "Email is required"
                },
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                }
            }
        },
...
    }
}